+++ eslint.config.mjs
@@ -1,0 +1,12 @@
+import unjs from 'eslint-config-unjs';
+
+export default unjs(
+  {
+    ignores: [
+      "test/fixture/jsonc/package.json"
+    ]
+  },
+  {
+    rules: {}
+  }
+);