+++ src/types/tsconfig.ts
@@ -4,22 +4,23 @@ export type StripEnums<T extends Record<string, any>> = {
   [K in keyof T]: T[K] extends boolean
     ? T[K]
     : T[K] extends string
-    ? T[K]
-    : T[K] extends object
-    ? T[K]
-    : T[K] extends Array<any>
-    ? T[K]
-    : T[K] extends undefined
-    ? undefined
-    : any;
+      ? T[K]
+      : T[K] extends object
+        ? T[K]
+        : T[K] extends Array<any>
+          ? T[K]
+          : T[K] extends undefined
+            ? undefined
+            : any;
 };
 
 export interface TSConfig {
   compilerOptions?: StripEnums<CompilerOptions>;
   exclude?: string[];
   compileOnSave?: boolean;
-  extends?: string;
+  extends?: string | string[];
   files?: string[];
   include?: string[];
   typeAcquisition?: TypeAcquisition;
+  references?: { path: string }[];
 }