+++ CHANGELOG.md
@@ -2,6 +2,181 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## v1.3.1
+
+[compare changes](https://github.com/unjs/pkg-types/compare/v1.3.0...v1.3.1)
+
+### 🩹 Fixes
+
+- Detect `bun.lock` ([#210](https://github.com/unjs/pkg-types/pull/210))
+
+### 🏡 Chore
+
+- Apply automated updates ([a006a0d](https://github.com/unjs/pkg-types/commit/a006a0d))
+- Update deps ([686c3aa](https://github.com/unjs/pkg-types/commit/686c3aa))
+- Update pathe to v2 ([edb9c09](https://github.com/unjs/pkg-types/commit/edb9c09))
+
+### ❤️ Contributors
+
+- Pooya Parsa ([@pi0](http://github.com/pi0))
+- Yacine Kanzari ([@yxuko](http://github.com/yxuko))
+
+## v1.3.0
+
+[compare changes](https://github.com/unjs/pkg-types/compare/v1.2.1...v1.3.0)
+
+### 🚀 Enhancements
+
+- **types:** Add `packageManager` field to `PackageJson` ([#204](https://github.com/unjs/pkg-types/pull/204))
+
+### 🏡 Chore
+
+- **release:** V1.2.1 ([395df23](https://github.com/unjs/pkg-types/commit/395df23))
+- Update deps ([cf98e6c](https://github.com/unjs/pkg-types/commit/cf98e6c))
+
+### 🤖 CI
+
+- Enable tests on windows ([#205](https://github.com/unjs/pkg-types/pull/205))
+
+### ❤️ Contributors
+
+- Pooya Parsa ([@pi0](http://github.com/pi0))
+- Colonel-Sandvich ([@Colonel-Sandvich](http://github.com/Colonel-Sandvich))
+
+## v1.2.1
+
+[compare changes](https://github.com/unjs/pkg-types/compare/v1.2.0...v1.2.1)
+
+### 🩹 Fixes
+
+- Optional fields from `publishConfig` ([#194](https://github.com/unjs/pkg-types/pull/194))
+- **tsconfig:** Add missing `references` key ([#198](https://github.com/unjs/pkg-types/pull/198))
+
+### 🏡 Chore
+
+- Update dependencies ([c173fc7](https://github.com/unjs/pkg-types/commit/c173fc7))
+- Lint ([ce3d808](https://github.com/unjs/pkg-types/commit/ce3d808))
+- Format packagejson.ts ([5cbff11](https://github.com/unjs/pkg-types/commit/5cbff11))
+- Apply automated updates ([e1234ea](https://github.com/unjs/pkg-types/commit/e1234ea))
+
+### ✅ Tests
+
+- **packagejson:** Add tests for partial objects ([e2dc735](https://github.com/unjs/pkg-types/commit/e2dc735))
+
+### ❤️ Contributors
+
+- Pooya Parsa ([@pi0](http://github.com/pi0))
+- Johan Meester <johan.meester71@gmail.com>
+- Kevin Deng 三咲智子 <sxzz@sxzz.moe>
+
+## v1.2.0
+
+[compare changes](https://github.com/unjs/pkg-types/compare/v1.1.3...v1.2.0)
+
+### 🚀 Enhancements
+
+- Add more fields into package.json ([#190](https://github.com/unjs/pkg-types/pull/190))
+
+### 🏡 Chore
+
+- Ignore fixture for renovate ([24c8fce](https://github.com/unjs/pkg-types/commit/24c8fce))
+- Update deps ([b539054](https://github.com/unjs/pkg-types/commit/b539054))
+
+### ❤️ Contributors
+
+- Kevin Deng 三咲智子 <sxzz@sxzz.moe>
+- Pooya Parsa ([@pi0](http://github.com/pi0))
+- 三咲智子 Kevin Deng <sxzz@sxzz.moe>
+
+## v1.1.3
+
+[compare changes](https://github.com/unjs/pkg-types/compare/v1.1.2...v1.1.3)
+
+### 🩹 Fixes
+
+- Fix types for `PackageJson` exports ([#185](https://github.com/unjs/pkg-types/pull/185))
+
+### 🏡 Chore
+
+- Update tsconfig and enable type tests ([#184](https://github.com/unjs/pkg-types/pull/184))
+
+### 🤖 CI
+
+- Enable `--typecheck` for vitest step ([cacfa8e](https://github.com/unjs/pkg-types/commit/cacfa8e))
+
+### ❤️ Contributors
+
+- Pooya Parsa ([@pi0](http://github.com/pi0))
+
+## v1.1.2
+
+[compare changes](https://github.com/unjs/pkg-types/compare/v1.1.1...v1.1.2)
+
+### 🩹 Fixes
+
+- Expand the types of `exports` ([#183](https://github.com/unjs/pkg-types/pull/183))
+
+### 🏡 Chore
+
+- Update deps ([f62330a](https://github.com/unjs/pkg-types/commit/f62330a))
+
+### ❤️ Contributors
+
+- Pooya Parsa <pooya@pi0.io>
+- Kevin Deng 三咲智子 <sxzz@sxzz.moe>
+
+## v1.1.1
+
+[compare changes](https://github.com/unjs/pkg-types/compare/v1.1.0...v1.1.1)
+
+### 🩹 Fixes
+
+- Check for all lockfiles before ascending directories ([#179](https://github.com/unjs/pkg-types/pull/179))
+
+### 🏡 Chore
+
+- Update lockfile ([bfd1fca](https://github.com/unjs/pkg-types/commit/bfd1fca))
+
+### ❤️ Contributors
+
+- Pooya Parsa ([@pi0](http://github.com/pi0))
+- Werheng Zhang ([@werheng](http://github.com/werheng))
+
+## v1.1.0
+
+[compare changes](https://github.com/unjs/pkg-types/compare/v1.0.3...v1.1.0)
+
+### 🚀 Enhancements
+
+- Auto detect and preserve format ([#175](https://github.com/unjs/pkg-types/pull/175))
+- Try to parse `package.json` as jsonc ([#176](https://github.com/unjs/pkg-types/pull/176))
+
+### 🩹 Fixes
+
+- Add object syntax browser field type ([#145](https://github.com/unjs/pkg-types/pull/145))
+- **types:** Allow `extends` as array ([#151](https://github.com/unjs/pkg-types/pull/151))
+- Add `imports` field to the types ([#160](https://github.com/unjs/pkg-types/pull/160))
+
+### 💅 Refactors
+
+- Migrate from `jsonc-parser` to `unjs/confbox` ([#174](https://github.com/unjs/pkg-types/pull/174))
+
+### 📖 Documentation
+
+- Add missing jsdocs ([#171](https://github.com/unjs/pkg-types/pull/171))
+
+### 🏡 Chore
+
+- Update repo ([f289b3a](https://github.com/unjs/pkg-types/commit/f289b3a))
+- Add build command ([984edb6](https://github.com/unjs/pkg-types/commit/984edb6))
+- Remove `.eslintignore` ([f1c1705](https://github.com/unjs/pkg-types/commit/f1c1705))
+
+### ❤️ Contributors
+
+- Pooya Parsa ([@pi0](http://github.com/pi0))
+- Max ([@onmax](http://github.com/onmax))
+- Parbez <imranbarbhuiya.fsd@gmail.com>
+
 ## v1.0.3
 
 [compare changes](https://undefined/undefined/compare/v1.0.2...v1.0.3)